Removed "notification type" since WAGNI (yet)
We figured that doing without it was more like the simplest thing that
could possibly work.

Note that we edited the migration in place after rolling back.  This
would cause problems if anyone had downloaded the code from pozorvlak's
repo the other day and run it, but we're assuming they didn't (since
they wouldn't have got anything interesting out of it anyway.)