Prettifying the plantings page.
Added a couple of tiny methods on Crop, and delegated them across to
Planting, to easily display the default/first scientific name, and to
show the number of times a given crop's been planted.

Sadly, the tests randomly fail on my box saying "No route matches
{:controller=>"plantings", :action=>"show"}" which is just nonsense. I
have no idea why it's doing that, so I'm committing it anyway, in the
hopes that either someone can tell me what's up, or else it's just a
glitch on my dev box.