Minor tweaks, mostly to formatting
... because I'm picky ;)  I particularly wasn't keen on the search link
on the logged in homepage, which didn't really fit in there.  I also
made the link display on *all* profiles (not just your own), and changed
the way the headings/titles were formatted.

I also changed CanCan to not specify who can search nearby members, but
to skip it in the controller instead.  It just didn't really make sense
in ability.rb.