Use PostgreSQL for Travis CI test runs.
We've encountered a couple of production bugs due to using different
databases in test and production. We're using SQLite in the test
environment to ease setup for newcomers, which fits with our broader
goals; this is an attempt to see if we can test on Postgres at the CI
stage to catch such bugs anyway.