Attempt at writing tests for planting reminder email
(Test are broken)

OK, so I decided to use the capybara_email gem to write these tests.
It's basically working but there's a problem with the has_link matcher.
Basically the emails sent by ActionMailer have URLs like
http://localhost:8080... whereas the tests asking for planting_url() or
similar are looking for http://example.com...

I don't understand AT ALL why there is this discrepancy, but it looks
like ActionMailer is using the settings from the development
environment, instead of the testing environment.

WHYYYYYY???

I'm pushing this up so that maybe someone else can figure it out,
because I'm stumped.