Improved layout of member profile page
In doing this, I added some content areas for all pages on the site:

1) subtitle
2) buttonbar

These are intended to help standardise the layout of all pages.  On the
member page, the subtitle is the location, and the buttonbar has links
like "edit profile", "upgrade account", etc (or if looking at someone
else's page, then "send message" etc).

I also implemented subtitle/buttonbar on the crop detail page (the
subtitle is the default scientific name).

The rest is just refactoring and tests.  I've removed some view tests
and put them in feature tests instead.