Hardcode file list in vendored active_utils gem
The active_utils gem generates its list of included files from git; this
means that if we run `bundle install` on a machine that doesn't have git
installed (for instance, a production or staging server), we get
annoying "fatal: Not a git repository" errors. This happens when we
deploy to staging or production - see, for instance,
https://travis-ci.org/Growstuff/growstuff/jobs/135737062#L2135.

This patch simply hardcodes the list of files to prevent these errors;
if we upgrade active_utils, we'll need to update the list.