Precompile assets in test
According to
https://github.com/teampoltergeist/poltergeist/issues/677#issuecomment-222919584,
a possible cause of the "Request failed to reach server" problem (#901)
is that assets are being compiled on-demand, causing enough of a
slowdown for PhantomJS to think the connection has failed. This turns
off lazy asset compilation in the test environment and precompiles
assets before testing on Travis. We should also turn off lazy asset
compilation in production - see
http://stackoverflow.com/questions/8821864/config-assets-compile-true-in-rails-production-why-not
Our Heroku deployments already run `rake assets:precompile`.